+++ ginger.py
@@ -36,7 +36,7 @@ Example:
   `ginger -f conllu input.conll -t tikz output.tex`
 """
 
-__version__ = 'ginger 0.5.1'
+__version__ = 'ginger 0.6.0'
 
 import sys
 import contextlib