+++ CHANGELOG.md
@@ -6,7 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-[Unreleased]: https://github.com/LoicGrobol/ginger/compare/v0.5.1...HEAD
+[Unreleased]: https://github.com/LoicGrobol/ginger/compare/v0.6.0...HEAD
+
+## [0.6.0] - 2017-04-03
+[0.6.0]: https://github.com/LoicGrobol/ginger/compare/v0.5.1...v0.6.0
 ### Fixed
   - Issue loading `package.json` in `setup.py` due to encoding
   - Force UTF-8 encoding on I/O